README.md :: added main.py :: commented galaina/R_code :: latest versions of Galaina R files plus script in preparation of new arrow assignment - dumb_script_for_testing_arc_direction.R - and draft causal graph exporter - just_for_playing_with_graph.R

+++ session/session.py
@@ -21,7 +21,8 @@ class Session:
         return self.file_pointer.read()
 
     def open_log(self):
-        os.remove(self.log_path)
+        if os.path.exists(self.log_path):
+            os.remove(self.log_path)
         file = open(self.log_path, 'a')
         file.close()
         self.file_pointer = open(self.log_path)
